# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Suivi Académique",
        "description": "Documentation de l'API pour le suivi académique",
        "contact": {
            "email": "contact@suivi-academique.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/salles": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Obtenir la liste des salles",
                "description": "Retourne la liste de toutes les salles",
                "operationId": "getSallesList",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "num_salle": {
                                                "type": "integer",
                                                "example": 101
                                            },
                                            "contenance": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "LIBRE"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Salles"
                ],
                "summary": "Créer une nouvelle salle",
                "description": "Crée une salle et retourne un message de confirmation",
                "operationId": "storeSalle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "num_salle",
                                    "contenance",
                                    "status"
                                ],
                                "properties": {
                                    "num_salle": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "contenance": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "OCCUPEE",
                                            "LIBRE"
                                        ],
                                        "example": "LIBRE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salle créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Salle créée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/salles/{num_salle}": {
            "get": {
                "tags": [
                    "Salles"
                ],
                "summary": "Afficher une salle spécifique",
                "description": "Retourne les détails d'une salle",
                "operationId": "getSalleByNum",
                "parameters": [
                    {
                        "name": "num_salle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "num_salle": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "contenance": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "LIBRE"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Salles"
                ],
                "summary": "Mettre à jour une salle",
                "description": "Met à jour les informations d'une salle",
                "operationId": "updateSalle",
                "parameters": [
                    {
                        "name": "num_salle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenance": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "OCCUPEE",
                                            "LIBRE"
                                        ],
                                        "example": "OCCUPEE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salle mise à jour avec succès"
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salles"
                ],
                "summary": "Supprimer une salle",
                "description": "Supprime une salle par son numéro",
                "operationId": "deleteSalle",
                "parameters": [
                    {
                        "name": "num_salle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle supprimée avec succès"
                    },
                    "404": {
                        "description": "Salle non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}